backend/requirements.txt
fastapi
uvicorn[standard]

pydantic
python-dotenv
requests

langchain-core
langchain-community
langgraph

typing-extensions
